src/requirements.txt
Django
psycopg2-binary
Pillow==5.4.1
requests==2.21.0
twilio==6.23.1
gunicorn
django-debug-toolbar==1.11